+++ Source/Graphics/Vulkan/VulkanGraphicsInterface.cpp
@@ -259,8 +259,8 @@ void VulkanGraphicsInterface::CreateInstance()
 	vkEnumerateInstanceLayerProperties(&layerCount, availableLayers.GetData());
 
 	VkDebugReportCallbackCreateInfoEXT debugInfo = Vk::DebugReportCallbackInfo(VulkanDebugCallback, debugFlags, this);
-	VkInstanceCreateInfo instanceInfo = Vk::InstanceInfo(validationLayers, ArraySize(validationLayers),
-		instanceExtensions, ArraySize(instanceExtensions), &debugInfo);
+	VkInstanceCreateInfo instanceInfo = Vk::InstanceInfo(validationLayers, (uint32)ArraySize(validationLayers),
+		instanceExtensions, (uint32)ArraySize(instanceExtensions), &debugInfo);
 	[[maybe_unused]] VkResult result = vkCreateInstance(&instanceInfo, nullptr, &instance);
 	CHECK_VK(result);
